oss-knowledge-intent-classifier/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
openai>=1.0.0
pyyaml==6.0.3
nest-asyncio==1.6.0
